backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
earthengine-api==0.1.393
osmnx==1.9.1
geopandas==0.14.3
shapely==2.0.3
scikit-learn==1.4.1.post1
httpx>=0.27.0
pvlib>=0.10.4
pandas>=2.0.0
google-genai==0.3.0
scikit-learn>=1.3.0
joblib>=1.3.2
python-dotenv>=1.0.0
